Enable software concurrent scavenge on z
1. Implement concurrent scavenge (CS) on z using ardbari opCode to generate
either software-based or hardware-based CS for reference field loads,
reference array copy, and object compare-and-swap.

2. Move referenceArrayCopy from OMR to this project.

3. Rename jitReadBarrier to softwareReadBarrier
and rename handleSoftwareReadBarrier to jitSoftwareReadBarrier on all
platforms

4. Update copyright info on files

5. Update comments where applicable

6. Use guarded-load-and-store loop to copy reference arrays when concurrent
scavenge is active. This applies to both hardware and software CS.

7. Change TR_J9VMBase::thisThreadGetEvacuateBaseAddressOffset() and
TR_J9VMBase::thisThreadGetEvacuateTopAddressOffset() to return
compressed range and uncompressed ranges on Z. This makes it possible for
z to perform compressed range check and can save
compression/decompression shift instructions. Hence, a faster
load-and-store loop.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>