Catch Exceptions during relocation
There are scenarios where an exception could be thrown during
relocation. For example, the SVM_ASSERT, in production builds, results
in a call to failCompilation (which throws an exception). Therefore,
this commit adds a try/catch around the call to applyRelocations to
ensure that the standard cleanup code for AOT load failures runs.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>