Get special-case invokeinterface receiver early
In 7295e2c8 the JIT compiler was modified so that it could generate IL
for "improper" invokeinterface bytecode instructions, i.e. those
targeting private methods or methods of Object. For these invocations,
the dispatch does not naturally perform the type check required for
invokeinterface, so an explicit type check is generated ahead of the
call. This type check needs the receiver of the call, which until now
has been read from the resulting call node. The problem is that it's
possible that the node is no longer a call by the time the type check is
generated, e.g. in the case of a call to Object.getClass(), the call
becomes aloadi <javaLangClassFromClass> during genInvokeDirect().

The receiver is now read from the stack before calling genInvokeDirect()
or genInvokeWithVFTChild() so that the type check will be unaffected by
any such transformations.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>