Revert "Revert "Look for additional opportunities for stack allocation""
This reverts commit cc13d295f5be4a2e8a8440d6a301a9fbad7fa775,
restoring commit c2eb8ed7f2ebc2f9d45d6d7c852027f0843747c1.

Repeating the log message for that restored commit:

This introduces some further refinements to detect opportunities for
Escape Analysis (EA) in loops where references are copied between local
variables.

This change considers the right-hand side of an assignment to be
harmless for the purposes of EA if it was an object that was locally
allocated by a previous pass of EA, not only if it is a candidate for
local allocation in the current pass of EA.

This change was prototyped by Andrew Craik <ajcraik@ca.ibm.com>.

Also, fixed a bug in trace code that was getting the class name of a
TR::New operation incorrectly.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>