Fix JNI check to properly handle atomic-free
When atomic-free was introduced, JNI check was modified to enter/exit
the VM instead of acquire/release VM access. This is incorrect in some
cases (particularly, the method exit hook which is called from inside
the VM with VM access).

The end result is that JNI check might incorrectly mark a thread as
being in a native when it is in fact running in the VM, resulting in
assertions or crashes.

Fix JNI check to differentiate between being called from JNI code, or
called internally from the VM.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>