Utilize fragmented memory as survivor in balanced GC
Currently region-base Balanced GC use MemoryPoolBumpPointer to manage
free memory, so both mutators and collectors only can make linear
allocation(only last free chunk in the region can be allocated), so some
of reclaimed free memory(the free chunks in the middle of region, such
as reclaimed free memory from GMP) can not be reused until the region
has been compacted or copyforward.

this change set is trying to utilize the "fragmented" memory as survivor
memory during copyforward collection in order to reduce pgc pause time,
copyforward abort cases and GMP collection count, then improve java
application runtime throughput and low-latency under balanced GC mode.

- mutators still allocate memory from eden regions, only collectors can
allocate from "fragmented free memory" as survivor memory.
- replace MemoryPoolBumpPointer with MemoryPoolAddressOrderedList in
order to manager the freelist for collector to allocate the free entry
in middle of the region.

- update SweepPoolManagerVLHGC extend from
SweepPoolManagerAddressOrderedList for building the freelist during
sweep and clear card table for free entries(only for the sweep during
first PGC after GMP, there is no compaction after sweep).

- free memory in survivor regions(not for empty region) need to be
aligned with CARD_SIZE before to be used by collectors for avoid
conflict for dirty card cases. the alignment process is incremental
along with collector allocation to avoid long pause for special case.
lostBytes for the alignment are marked as darkmatter.

- clearing card table for free memory is piggybacked by building
freelist during sweep, and lostBytes due to the card alignment is also
calculated(the alignment process is postponed) for avoid wrong free
memory report for the estimation.

- compressed survivor card table( 1bit for 1 CARD_SIZE/512 Bytes heap
memory) is maintained during copyforward for distinguishing if live
object in survivor memory in a non empty survivor region.

- compressed survivor card table is cleared at the begin of copyforward
and the related bits are updated when the free memory is reserved for
object or copycache.

- currently any discarded partial copy cache is treated as a darkmatter.

- new XXgc:minimumFreeSizeForSurvivor option for setting minimum Free
Entry can be reused as survivor memory(default = 2048).

- related DDR change

Signed-off-by: Lin Hu <linhu@ca.ibm.com>