Fix up global continuation lists instead of rebuilding for compaction
Once continuation Object is constructed, the Object would be added in
global continuation lists and the "dead" Object would be collected by
GC, since logically compaction would not collect any continuation Object
in the list, so we piggyback heap fix up to rebuild the list to reduce
the extra cycle to walk the old list for fixing up.
But between the Object is created and added into the list, GC could be
triggered and find the live Object, which has not be added in the list,
it could cause a circular linklist case, so changing fix up the existing
list instead of rebuilding the list for compaction.

Signed-off-by: Lin Hu <linhu@ca.ibm.com>