Fix wrong memoryPool names of memory usage in GCInfo
GCInfo contains a set of memory usages for all of memoryPools(both Heap
and NonHeap),java function buildGCInfo() retrieve a list of MemoryPools'
names and a list of Usages internally and generate the set of memory
usages, in some platforms(ex. MacOS) the order for the list of names and
for the list of usages might be different, need to pass memorypool id to
match the order between the names and the usage.

also update to use different bit to present heap memory pools and non
heap memory pools for the internal memorypool ID, it would avoid
pontetial confusion.

Signed-off-by: hulin <linhu@ca.ibm.com>